backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
sgp4==2.22
scipy==1.11.2
huggingface-hub==0.19.4
markdown==3.5.1
python-dotenv==1.0.0
huggingface-hub==0.19.4
markdown==3.5.1
python-dotenv==1.0.0